Rename http_client to _http_client; #44
This allows _http_client to be monkey-patched even while
balanced.http_client is assigned an instance of HTTPClient.
